Revert latest changes to atom.xml feed
This reverts commit a18b2095bc1efd229f8366e69bac01aad6d6bc49.

The latest change to the atom.xml file introduced duplicate entries, and
also an invalid entry for "latest". You don't need a latest in the atom
feed, porter calculates it based on timestamps and version numbers.

Unfortunately the latest entry has bad URLs, so the porter client is
respecting them and then can't download the file.

The fix is to revert the changes to the file, and educate the user who
was originally reporting the error on how to install this mixin.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>